refactor: adopt kubebuilder v4 project layout
Move main.go to cmd/main.go and controllers/ to internal/controller/ (package controllers -> controller), bump PROJECT layout to go.kubebuilder.io/v4, and update the Makefile build/run paths, goreleaser (main: ./cmd), the envtest CRD path (now two levels deep), and the IDE run configs. Collapse the duplicate ENVTEST_K8S_VERSION (now 1.33.0).
