+++ .goreleaser.yaml
@@ -1,6 +1,7 @@
 project_name: gitlab-runner-operator
 builds:
   - env: [CGO_ENABLED=0]
+    main: ./cmd
     binary: operator
     goos:
       - linux
