+++ .run/Controller.run.xml
@@ -1,13 +1,13 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Controller" type="GoTestRunConfiguration" factoryName="Go Test">
     <module name="gitlab-runner-operator" />
-    <working_directory value="$PROJECT_DIR$/controllers" />
+    <working_directory value="$PROJECT_DIR$/internal/controller" />
     <envs>
       <env name="TEST_TIMEOUT" value="35" />
       <env name="USE_EXISTING_CLUSTER" value="true" />
     </envs>
     <kind value="PACKAGE" />
-    <package value="gitlab.k8s.alekc.dev/controllers" />
+    <package value="gitlab.k8s.alekc.dev/internal/controller" />
     <directory value="$PROJECT_DIR$" />
     <filePath value="$PROJECT_DIR$" />
     <framework value="gotest" />
