+++ .run/RunnerControllerTest.run.xml
@@ -1,10 +1,10 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="RunnerControllerTest" type="GoTestRunConfiguration" factoryName="Go Test">
     <module name="gitlab-runner-operator" />
-    <working_directory value="$PROJECT_DIR$/controllers" />
+    <working_directory value="$PROJECT_DIR$/internal/controller" />
     <root_directory value="$PROJECT_DIR$" />
     <kind value="PACKAGE" />
-    <package value="gitlab.k8s.alekc.dev/controllers" />
+    <package value="gitlab.k8s.alekc.dev/internal/controller" />
     <directory value="$PROJECT_DIR$" />
     <filePath value="$PROJECT_DIR$" />
     <framework value="gotest" />
