+++ internal/controller/multirunner_controller_test.go
@@ -1,4 +1,4 @@
-package controllers
+package controller
 
 //
 // import (
